Add IP-based calculation regardless of permissions.
When you load the page, it asks if you want to share your location. In
Safari, there are three options, yes, never, and not now. If you choose
not now, nothing happens. The callback in case of failure is not called.
This is by design - apparently people ignore that question a lot and web
developers need to handle that case. There are other ways we could
handle it, but this is the simplest. The downside is that if the user
says yes (in any browser), one number will show and then, if the
geoposition-based number is different, the number will change.