Catch more bad interpolation outcomes.
Negative number (imagine you're south of three points
where the northernmost is getting some snow and the southernmost
is getting none, the plane will say you should get less than none)
gets reported as zero. NAN (if the points are in a line, probably
doesn't happen but just in case) gets reported as out of range.